+++ R/show_ranef.R
@@ -13,8 +13,8 @@
 #'
 #' @author Steven V. Miller
 #'
-#' @param data a fitted mixed model with random intercepts
-#' @param grp What random intercept/slopes do you want to see as a caterpillar plot? Declare it as a character
+#' @param model a fitted mixed model with random intercepts
+#' @param group What random intercept/slopes do you want to see as a caterpillar plot? Declare it as a character
 #' @param reorder optional argument. DEFAULT is TRUE, which ``re-orders'' the intercepts by the
 #' original value in the data. If FALSE, the ensuing caterpillar plot defaults to a default method of ordering
 #' the levels of the random effect by their estimated conditional mode.
@@ -29,21 +29,41 @@
 #' show_ranef(M1, "Subject")
 #' show_ranef(M1, "Subject", reorder=FALSE)
 #'
+#
+# show_ranef <- function(data, grp, reorder = TRUE) {
+#   data <- augment(ranef(data, condVar = TRUE))
+#   if (reorder) {
+#     data <- data[data$grp == grp, ]
+#     data$level <- as.character(data$level)
+#   } else {
+#     }
+#   ggplot(data[data$grp == grp, ], aes(.data$estimate, .data$level, xmin = .data$lb, xmax = .data$ub)) +
+#     geom_errorbarh(height = 0) +
+#     geom_vline(xintercept = 0, lty = 2) +
+#     geom_point() +
+#     ggplot2::facet_wrap(~variable, scale = "free_x") +
+#     ylab("Levels of the Random Effect") +
+#     xlab("Estimated Intercept")
+# }
 
-show_ranef <- function(data, grp, reorder = TRUE) {
-  data <- augment(ranef(data, condVar = TRUE))
+# Something to look at for later: https://fishandwhistle.net/slides/rstudioconf2020/#1
+
+# Let's try this again after BDR's email from March 22
+
+show_ranef <- function(model, group, reorder = TRUE) {
+  data <- data.frame(ranef(model, condVar= TRUE))
+  data$ub <- data$condval + ((p_z(.1)*data$condsd))
+  data$lb <- data$condval - ((p_z(.1)*data$condsd))
   if (reorder) {
-    data <- data[data$grp == grp, ]
-    data$level <- as.character(data$level)
+    data <- data[data$grpvar == group, ]
+    data$grp <- as.character(data$grp)
   } else {
-    }
-  ggplot(data[data$grp == grp, ], aes(.data$estimate, .data$level, xmin = .data$lb, xmax = .data$ub)) +
+  }
+  ggplot(data[data$grpvar == group, ], aes(.data$condval, .data$grp, xmin = .data$lb, xmax = .data$ub)) +
     geom_errorbarh(height = 0) +
     geom_vline(xintercept = 0, lty = 2) +
     geom_point() +
-    ggplot2::facet_wrap(~variable, scale = "free_x") +
+    ggplot2::facet_wrap(~term, scale = "free_x") +
     ylab("Levels of the Random Effect") +
     xlab("Estimated Intercept")
 }
-
-# Something to look at for later: https://fishandwhistle.net/slides/rstudioconf2020/#1