+++ NEWS.md
@@ -2,6 +2,7 @@
 
 - Add `filter_refs()` and, with it, the `{bib2df}` package as a dependency.
 - `print_refs()` will now work on an (implied) `{bib2df}` data frame of `.bib` entries.
+- Add `wom()`.
 
 # stevemisc 1.3.0